be 배포테스트 ouath and dairy

+++ server/src/main/java/com/pathfinder/server/diary/service/DiaryService.java
@@ -1,20 +1,26 @@
 package com.pathfinder.server.diary.service;
 
+import com.pathfinder.server.auth.utils.SecurityUtil;
+import com.pathfinder.server.diary.entity.Diary;
+import com.pathfinder.server.diary.repository.DiaryRepository;
+import com.pathfinder.server.global.exception.diaryexception.DiaryDeleteUnAuthorizedException;
+import com.pathfinder.server.global.exception.diaryexception.DiaryEditUnAuthorizedException;
+import com.pathfinder.server.global.exception.diaryexception.DiaryNotFoundException;
 import com.pathfinder.server.member.entity.Member;
 import com.pathfinder.server.member.service.MemberService;
-import com.pathfinder.server.diary.entity.Diary;
-import com.pathfinder.server.exception.BusinessLogicException;
-import com.pathfinder.server.exception.ExceptionCode;
-import com.pathfinder.server.diary.repository.DiaryRepository;
 import com.pathfinder.server.reward.service.RewardService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 @Service
+@Transactional
 public class DiaryService {
     private final DiaryRepository diaryRepository;
     private final MemberService memberService;
@@ -32,19 +38,26 @@ public class DiaryService {
         return diaryRepository.save(diary);
     }
 
+    @Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.SERIALIZABLE)
     public Diary updateDiary(Diary diary){
         Diary findDiary = findVerifiedDiary(diary.getDiaryId());
-        Optional.ofNullable(diary.getTitle())
-                .ifPresent(title -> findDiary.setTitle(title));
-        Optional.ofNullable(diary.getContent())
-                .ifPresent(content -> findDiary.setContent(content));
-        Optional.ofNullable(diary.getArea1())
-                .ifPresent(area1 -> findDiary.setArea1(area1));
-        Optional.ofNullable(diary.getArea2())
-                .ifPresent(area2 -> findDiary.setArea2(area2));
-        return diaryRepository.save(findDiary);
+        if(verifyIdentification(findDiary)){
+            Optional.ofNullable(diary.getTitle())
+                    .ifPresent(title -> findDiary.setTitle(title));
+            Optional.ofNullable(diary.getContent())
+                    .ifPresent(content -> findDiary.setContent(content));
+            Optional.ofNullable(diary.getArea1())
+                    .ifPresent(area1 -> findDiary.setArea1(area1));
+            Optional.ofNullable(diary.getArea2())
+                    .ifPresent(area2 -> findDiary.setArea2(area2));
+            return diaryRepository.save(findDiary);
+        }
+        else {
+            throw new DiaryEditUnAuthorizedException();
+        }
     }
 
+    @Transactional(readOnly = true)
     public Diary getDiary(Long diaryId){
         Diary findDiary = findVerifiedDiary(diaryId);
         findDiary.setViews(findDiary.getViews() + 1); // 조회수 증가
@@ -71,13 +84,19 @@ public class DiaryService {
 
     public void deleteDiary(Long diaryId) {
         Diary findDiary = findVerifiedDiary(diaryId);
-
-        diaryRepository.delete(findDiary);
+        if(verifyIdentification(findDiary)){
+            diaryRepository.delete(findDiary);
+        }
+        else {
+            throw new DiaryDeleteUnAuthorizedException();
+        }
     }
 
+
+    @Transactional(readOnly = true)
     public Diary findVerifiedDiary(Long diaryId) {
         Optional<Diary> optionalQuestion = diaryRepository.findById(diaryId);
-        Diary findDiary = optionalQuestion.orElseThrow(()-> new BusinessLogicException(ExceptionCode.DIARY_NOT_FOUND));
+        Diary findDiary = optionalQuestion.orElseThrow(()-> new DiaryNotFoundException());
         return findDiary;
     }
     private void verifyDiaryGetMemberName(Diary diary){
@@ -85,5 +104,11 @@ public class DiaryService {
         diary.setName(findUser.getName());
         findUser.setDiaryCount(findUser.getDiaryCount() + 1);
         rewardService.unlockRewards(findUser,findUser.getRewards());
+    }
+    public boolean verifyIdentification(Diary diary) {
+        if(diary.getMember().getMemberId() == SecurityUtil.getCurrentId()) {
+            return true;
+        }
+        return false;
     }
 }
